backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
sqlalchemy>=2.0,<3.0
pydantic>=2.10,<3.0
pillow>=11.0,<12.0
numpy>=1.26,<3.0
python-multipart>=0.0.17,<1.0
python-dotenv>=1.0,<2.0
pymysql>=1.1,<2.0
cryptography>=43,<46
alembic>=1.13,<2.0
# v0.5 optional - crawler & async tasks
redis>=5.0,<6.0
arq>=0.26,<1.0
playwright>=1.48,<2.0
# v0.5 optional - local background removal
rembg>=2.0,<3.0
# v0.5 optional - knowledge base (RAG)
qdrant-client>=1.12,<2.0
sentence-transformers>=3.0,<4.0
